Change value functions to support 2-way bindings, remove old conditional bindings.
Changes: A node may have different value functions depending on which
node it receives a value from. Conditions were previously specified
using conditional bindings, with the order of declaration specifying
the order in which the conditions were evaluated, this has been
removed and is to be replaced, with explicit conditional constructs,
in a future commit.