Change parser to take operator precedence into account when parsing function application.
Function application is also treated as having a precedence, which is
taken into account when parsing expressions. This allows the
precedence of certain operators to be set higher than the precedence
of function application.

E.g. if the dot '.' operator has a higher precedence than function
application:

m.fn(a, b)

will be parsed as ((. m fn) a b) and not (. m (fn a b)).