Fix bugs in builder.
Since the global meta-nodes table was being copied into each
meta-node's definition an infinite loop was caused when building the
meta-node definitions.

Referencing outer-nodes, from inside a meta-node, would always result
in an error since add-node was being used, which adds a new node and
raises an error if a node with the same name already exists, instead
of ensure-node.