Add more html file builder tests and fix issue.
The tests test that referencing the `self`, which is an alias for the
HTML component node, from a meta-node defined inside an HTML file,
work correctly.

Fixed the issue with an error being raised when the module, in which the
HTML component node is to be created, does not yet exist.