Add error checking and fix issues.
Error checking is added ensuring that module pseudo-nodes do not
appear as the source of a binding or as operands, that anything
that is not an ordinary node does not appear as the target of a
binding, and that only nodes appears as functor operators.

The issue with referencing meta-nodes as the source of a binding has
been fixed.