Change node auto-creation and referencing semantics.
Nodes defined in an outer scope can be referenced from within
meta-nodes directly, without the `..` operator. This comes with a
restriction on when local nodes are automatically created. Local nodes
are automatically created only when they appear in the target position
of a binding. There are no changes to node auto-creation in the global
scope.

Additionally nodes, meta-nodes and module pseudo-nodes are all in a
single namespace (symbol table) which simplifies the implementation
and semantics.