Add basic strictness analysis.
Use the strictness analyzer to avoid compiling some meta-node
arguments to `THUNK's. This allows proper tail call optimization to be
performed whereas without strictness analysis, the stack of `THUNK's
grows even when the function call stack does not.

Additionally the expression in tail position is always returned as a
thunk, which allows for general tail call optimization.