Change evaluation strategy in JS backend.
Meta-nodes do not always return thunks, rather expression which are
conditionally evaluated, such as the then and else branches of if
expressions, are always wrapped in thunks.

Fix issue with failing expressions in JS backend.  Each expression,
which explicitly resolves at least one of its arguments, is surrounded
in a try-catch block, in order for any failures to be passed on to the
parent expressions.