Fix issue with external meta-nodes with optional arguments.
Previously the meta-nodes had no way of knowing whether the optional
arguments were supplied or not other than an adhoc check for NIL. NIL
however may also be a boolean return value, thus is not a good way.

The fix is to not pass NIL for the values of optional arguments with
no default values, when invoking an external meta-node function.