Add functionality for querying failure types.
A `fail-type` meta-node is added which, if its argument fails to
evaluate to a value, returns the failure type.

The `:context` operator is enhanced to take a `test`
function/meta-node as its third argument. If the context's previous
value function failed to evaluate to a value, `test` is applied on the
failure type. If `test` returns true the binding takes effect
otherwise the next binding is tested.