Add new meta-node classes to distinguish between build stages.
`META-NODE-SPEC' represents a meta-node before it has been built. It
contains the operand node names and declarations comprising the body.

`BUILT-META-NODE' represents a meta-node, of which the definition has
been built however the final build stages have not yet been performed.

`FINAL-META-NODE' represents a meta-node which has been built and all
the final build stages have been performed. The outer node set of a
meta-node in this build stage is complete.