Add generation of referencing loading and storing instructions.
Local variables are stored on the stack in order for the root
references to be visible to the garbage collector.

A garbage collection may occur during any memory allocation, after
which the root references may be changed. The references which are
stored in the local variables are thus stale and have to be reloaded
from the stack.

The load and store instructions are not generated if it can be proven
that no garbage collection has taken place since the last refresh.