Fix issue with previous node value references in Wasm backend.
The compute function is passed two state arrays, the previous state
array and a copy of it in which the new node state will be stored. The
old state array is not modified.

Previously the old state array was modified, with the new values of
the input nodes. This caused issues if a thunk, which reference the
previous values of nodes which depend on theses input nodes, is
resolved after the state update.