Fix major bug with constant folding in Wasm backend.
Folding a constant into multiple nodes, or into the bodies of multiple
meta-nodes, resulted in EXPRESSION-BLOCK objects, contained in the
value expression of the constant node, being shared across multiple
context/meta-node functions. This itself, resulted in the usage counts
of the EXPRESSION-BLOCK's being greater than the number of times they
are actually used in the respective context/meta-node function.