Fix stale pointer issues in wasm runtime library.
When a memory allocation takes place, all used pointers are saved on
the stack prior to calling `alloc` and restored from the stack after
`alloc` returns. This is done since each call to `alloc` may trigger a
garbage collection, thus rendering the pointers invalid.

Additionally fix issue with lazy evaluation and marshalling of objects
from Wasm to JavaScript:

All thunks are computed when an object is being marshalled from the
WebAssembly heap to JavaScript in order to avoid stale pointer issue
if the computation of a thunk triggers a garbage collection
cycle. This however means that lazy evaluation is only supported
within Tridash itself and not when marshalling values from Tridash to
JavaScript.